Real time corner case fix
It works, but not great yet - need to figure out the right way to delay the decision for the type of peak that is being evaluated second. Probably not using PreVARP/PostVARP/PreAVRP/PostAVRP correctly. We should also consider consolidating the huge if/else statement in yamPeakFinder into one block, as it is repetitive and unweildy, but doing so would require big structural changes.